Update patch set 1
Patch Set 1: I would prefer that you didn't merge this

Actually, the intent was to have an easy way to output a single job to stdout, but it looks like I accidentally lost that along the way.  I'd prefer to either leave -o optional and fix it so that you can still run:

 jenkins-jobs test config job-name

and have that output the xml to stdout, or perhaps allow "-" as an option to "-o" that would output to stdout.

Patch-set: 1
Label: Verified=0
Label: Code-Review=-1
Label: Workflow=0