Update patch set 2
Patch Set 2:

The request changes make sense. 
(I originally stuck in the logging statements because I was having difficulty getting the correct structure in my YAML files.  I was getting errors from the gen_xml() functions but I didn't know which file was the culprit.  Perhaps some kind of schema verification on the YAML would help with this, but I dunno if it'd be worth the effort (for me), now that I have a handle on the YAML format.)

Patch-set: 2