Update patch set 6
Patch Set 6: I would prefer that you didn't merge this

Two things:

1) Can you add inline documentation?  See the other modules for an example.  The docs are new since you first proposed this.

2) Consider this alternate implementation, which I think will be a little cleaner, and should get some duplicated logic (like iterating over all jobs) out of the module:

Don't use handle_data, use gen_xml instead, in which you create or add to the properties element as needed (see the 'parameters' module for an example of that), and the same for the 'publishers' property (this is also new since you first proposed this patch).

Since gen_xml is already per-job, that gets rid of that bit of complexity.  You also don't need any property or publisher functions, which is fine and makes more sense because the hipchat configuration (split over the two as it is) doesn't make sense as either of those.

Patch-set: 6
Label: Verified=0
Label: Code-Review=-1
Label: Workflow=0