Update patch set 1
Patch Set 1: I would prefer that you didn't merge this

We don't want to require that everyone install the hipchat python module if they aren't going to use it, so it will need to fail gracefully if the import fails (and then probably raise an exception if someone then tries to use it in the yaml).

I think we should pass the config object into the ModuleRegistry, and then the hipchat module can access it via the registry which is passed into the module's __init__ method.

Also, I think you should move the hipchat property function to the hipchat module -- it doesn't have to live in the same module as the other properties, the entry point can just point to it directly in the hipchat module.  that will make for better code organization, I think.

Patch-set: 1
Label: Verified=0
Label: Code-Review=-1
Label: Workflow=0