Update patch set 1
Patch Set 1: Do not merge

I tested this on jenkins-dev.o.o, which is running 1.482 (we recently ugraded to patch a critical security flaw in Jenkins), and this failed to cause the envinject section of the images-api-v20 job to be created.  I double checked, and HEAD is still able to do that.  So, this doesn't seem to work in 1.482.

Patch-set: 1
Label: Verified=0
Label: Code-Review=-2
Label: Workflow=0