Update patch set 2
Patch Set 2:

Whaou thanks for the kind comments :-)

PS2 address the documentation issues:
* link to the Ant plugin. Apparently shipped by default but better safe than sorry.
* properly document parameters using:  args: <type> <name> <comment>
* fixed the yaml syntax to use a dictionary relying on iteritems() to examine the parameters.

As for polymorphism, I basically found out it is more concise to write:

 - ant: "target1 target2

rather than:

 -ant:
   targets: "target 1 target2"

It is probably disruptive and not in par with the rest of the builder. If it ends up being confusing to the users, I will be more than happy to remove that shortcut and update the patchset to only support the second form using explicit parameters.

Patch-set: 2