Update patch set 1
Patch Set 1: I would prefer that you didn't merge this

This is great!  I have a few comments:

(+1 to Clark's note about documenting the module dependency.)

Polymorphism: I don't believe there are any polymorphic
components.  It's a clever idea, and pretty elegant, but since
YAML parsing is so picky, and validation isn't very good, I think
consistency helps.  I'm inclined to say we shouldn't do
polymorphic forms.  I could be convinced otherwise if you feel
strongly, but if we do, then we should try to standardize it a
bit, including a common documentation format, and a mention of
the technique in the main documentation.

Parameters: Can you add parameter tags, like this:

  :arg str targets: the space separated list of ANT targets
  :arg str buildfile: the path to the ANT build file

That will make the nice table of parameters for the builder in
the documentation.  This is one of the things I think we'd need
to address if we do use polymorphism -- I'd like to see a table
for each alternate form.

Parameter type: Currently the parameters show up as one element
lists, but since nothing is being listed, I think they should be
a dictionary on the ant item.  So it should look like this:

  builders:
    - ant:
        targets: "debug test install"
        buildfile: "build.xml"

See the SCP publisher for an example.

Patch-set: 1
Label: Verified=0
Label: Code-Review=-1
Label: Workflow=0