Update patch set 1
Patch Set 1:

To David Ripton:
> There seems to be a case error, "thresholdmode" in most places versus "thresholdMode" in one place. (And "Thresholdmode" in a comment.)

That could indeed be a bit confusing. The mixed case one is used to forge an XML element that ends up in the Jenkins configuration. Since XML is case sensitive, we need to respect that mixed case.

>  I don't really like the constants '1' and '2' for thresholdmode. If you're using strings anyway, why not use descriptive strings like 'absolute' and 'relative'?

I to find them awkward, would have been simpler if the author used a string instead of int + constants. Since the plugin only recognize 1 or 2, we have to use that value in the XML configuration.

My first patch used 1/2, found them to be meaningless in a yaml file so I have made them 'percent' and 'number' and let the job builder to convert them.

Patch-set: 1
Label: Code-Review=0