Update patch set 1
Patch Set 1: Looks good to me (core reviewer)

The one thresholdmode that could be changed is in that comment. (the other seems to be about the xml element, so I'm ok with it) - so if you want to change that, go for it.

The 1 and 2 values are defined on the xunit plugin side of things and we have to give them those values as input, yeah? Seems like the yaml-side interface is number and percent, so I like that.

Patch-set: 1
Label: Code-Review=+2
Label: Workflow=0