Update patch set 1
Patch Set 1: Looks good to me, but someone else must approve

(1 inline comment)

Correct, the 1 and 2 values are what need to end up in the XML configuration for the xunit plugin.

There is one small change to the docstring that I think needs to be made, but other than that this looks good to me.

Patch-set: 1
Label: Code-Review=+1
Label: Workflow=0