(WIP) tests for publishers (WIP)
This is a work in progress.

I am trying to build up a very basic testing framework so we can tests
our publishers. The idea is to provide a YAML input file and expected
XML output, run the parser on the YAML file and compares its output with
the fixture XML.

That test suite should speed up developpement of new publishers.

Work in progress.

The discover plugin find the TestCaseModulePublisher class and attempt
to load it with only one parameter.  The test suite should rely on
load_tests() which would provide the test suite but discover is not
smart enough.

Issue:

 $ python -m discover
Traceback (most recent call last):
  File "2.7/lib/python2.7/runpy.py", line 162, in _run_module_as_main
    "__main__", fname, loader, pkg_name)
  File "2.7/lib/python2.7/runpy.py", line 72, in _run_code
    exec code in run_globals
  File "site-packages/discover.py", line 480, in <module>
    main()
  File "site-packages/discover.py", line 464, in main
    tests, verbosity = _do_discovery(argv, verbosity, testLoader)
  File "site-packages/discover.py", line 440, in _do_discovery
    return loader.discover(start_dir, pattern, top_level_dir), verbosity
  File "site-packages/discover.py", line 252, in discover
    tests = list(self._find_tests(start_dir, pattern))
  File "site-packages/discover.py", line 321, in _find_tests
    for test in self._find_tests(full_path, pattern):
  File "site-packages/discover.py", line 321, in _find_tests
    for test in self._find_tests(full_path, pattern):
  File "site-packages/discover.py", line 302, in _find_tests
    yield self.loadTestsFromModule(module)
  File "site-packages/discover.py", line 105, in loadTestsFromModule
    tests.append(self.loadTestsFromTestCase(obj))
  File "site-packages/discover.py", line 96, in loadTestsFromTestCase
    loaded_suite = self.suiteClass(map(testCaseClass, testCaseNames))
TypeError: __init__() takes exactly 4 arguments (2 given)

Change-Id: If3260113eb1337ac47c3883b11c600e5a595dae3