(WIP) tests for publishers (WIP)
This is a work in progress.

I am trying to build up a very basic testing framework so we can tests
our publishers. The idea is to provide a YAML input file and expected
XML output, run the parser on the YAML file and compares its output with
the fixture XML. Ideally should show a nice difference.

That test suite should speed up developpement of new publishers.

Work in progress.

Change-Id: If3260113eb1337ac47c3883b11c600e5a595dae3