Update patch set 1
Patch Set 1: I would prefer that you didn't merge this

Jeblair, I think this is related to the previous change that allows per job variable interpolation in the project definition. Instead of specifying which job to trigger in the job template because it may be different for each pipeline you define it in the project definition on a per job basis. But you need a clean way to kill that pipeline so the last job in the pipeline will get a variable value of empty string. This is not interpolated and instead signifies that no trigger/pipeline XML should be generated.

That said, I think adding something that explains this to the documentation is a good idea.

Patch-set: 1
Label: Code-Review=-1
Label: Workflow=0