Update patch set 1
Patch Set 1: I would prefer that you didn't merge this

The properties module of job-builder has support for envinject.  Of course, envinject can be used both as a job-level option (in properties) or a build step (in builders) as you've done here, so it's a good idea to round out the implementation by adding support for the builder.

However, I think that the name of the component should match ('inject' vs 'env-inject', and so should the options (at least, those that it has in common).  Can you change the name/options in the builder to match the current proprerties component?

Patch-set: 1
Label: Verified=0
Label: Code-Review=-1
Label: Workflow=0