Update patch set 4
Patch Set 4: I would prefer that you didn't merge this

I think the usage of ${FOO} is similar enough to the {substitution} elsewhere in job-builder that it warrants an explanation in the docstring.  I mean, no need to duplicate the documentation of the plugin or anything, but just explaining what ${FOO} is in the example would be helpful.  It confused me at first.

Patch-set: 4
Label: Verified=0
Label: Code-Review=-1
Label: Workflow=0