Update patch set 4
Patch Set 4: Looks good to me (core reviewer)

You would need a job template that used the build name setter to be passed a variable with a name of FOO to have Build-${FOO} do a string replacement. As long as the thing in ${} is not a variable known to jenkins job builder it should ignore it. Might be worth adding a note about that in the docstring, but this looks good to me.

Patch-set: 4
Label: Code-Review=+2
Label: Workflow=0