Update patch set 2
Patch Set 2: No score

Agree with clark -- The layout looks fine, but the additional info in the comments helps, so that should probably go into the docs too.

And yeah, in general, I like to approximate the web UI, rather than the XML, and keep it as close to the web ui as possible.  But also, JJB is an opportunity to improve configuration if there's something that works better a different way in YAML than in HTML.

Patch-set: 2
Label: Code-Review=0
Label: Workflow=0