Update patch set 1
Patch Set 1: Looks good to me (core reviewer)

LGTM. Would be nice if the first line of the commit message was a nice short summary. eg

Add ablity to use the Env Inject Plugin in the Build Environment section.

This change adds a third location for the use of the Env Inject plugin.

Patch-set: 1
Label: Code-Review=+2
Label: Workflow=0