Update patch set 1
Patch Set 1: I would prefer that you didn't merge this

Fix the pep8 issue and this lgtm. This doesn't seem so terribly bad as a work around to the issue. Every zuul job needs this notification url, so it is a global default in that way. Also, it isn't crazy horrible to understand what is going on there. tl;dr I can live with this.

Patch-set: 1
Label: Code-Review=-1
Label: Workflow=0