Update patch set 2
Patch Set 2:

No, unfortunately, I don't.  I was hoping you would be able to give me some pointers.  My thought was that builder.py gets called first so that is why the stuff showed up at the top like you would expect, and now that it is in the module toplevel.py it is getting called later and thus puts it in at a later moment.  I wasn't super happy that my change put the actions tag and description etc in the middle, it should be at the top like jenkins does it.

Patch-set: 2
Label: Code-Review=0