Update patch set 5
Patch Set 5:

Used  XmlJob() to prettify the XML. The tests now reports a unified diff of the XML from the fixture file versus the generated one and is currently failing.   need to find out how the various types are described in a real Jenkins. I guess there should be only one <types/> with everything in it.

Patch-set: 5