Update patch set 3
Patch Set 3: I would prefer that you didn't merge this

Ideally we'd maintain an ordering as close to jenkins as possible, but the only way to do that would be to have separate modules.  If we want to combine them for convenience, we won't be able to control the ordering as well.  That should be fine -- tests so far have indicated that the order doesn't affect Jenkins.

I'd like to see a few changes to make this a more comprehensive change:

1) Maybe call the top level module "general", to match existing documentation.

2) Include logrotate in the new general module.

3) Give it sequence 10 (logrotate's old sequence).

4) Be sure to include the documentation from assignednode.py and logrotate.py.

5) Change the documentation for the "General Job Config" page to use the module docs from #4.

6) Move the docs from http://ci.openstack.org/jenkins-job-builder/configuration.html#job into the module docs for the new general module.

7) Have the module docs in the general module use the style from this change:  https://review.openstack.org/#/c/16908/5/jenkins_jobs/modules/project_matrix.py

I know that's a lot more work than you set out to do with this change, but it's a little bit of a change in direction -- I'm supportive of it, though I want to make sure we're thorough and don't leave things half the old way, and half the new way.

Thanks!

Patch-set: 3