Update patch set 2
Patch Set 2: (1 inline comment)

I think it would be very nice for the description and actions and those things that Jenkins puts at the top to be at the top from this project.  The things inside builders or properties are as necessary, but the top level ones should probably be at the top of the xml output.

Patch-set: 2