Update patch set 7
Patch Set 7:

This is the same error the last few merged commits have gotten and I don't know if there is anything I can do to fix it, y'all might have to fix it on the server side?  If not what is the correct url for the clone of puppet, I can't find it anywhere.

Patch-set: 7