Fixing Inject to not create empty tags
Using the Inject functionality with the builders section caused
an error in Jenkins because the script-file was blank but the tag
existed in the xml.  Fixing this by only putting in the tag if
the value exists in the yaml.  This changes the inject functionality
in builders.py properties.by and wrappers.py.  It checks for the
following properties to exist before creating the tag: properties-file
properties-content script-file script-content and groovy-content

Change-Id: I9151f7444f6807e929721b04fb8e7ad3a867cd6c