Update patch set 1
Patch Set 1:

The properties-file being empty in the builders section was causing the build to fail because it couldn't find the file at ''.  When I just hit the save button on the job in jenkins it then removed the tag, which made me think this was the correct thing for JJB to do.  I went ahead and preemptively did the same thing for all instances of the env inject plugin.

Patch-set: 1
Label: Code-Review=0