Update patch set 2
Patch Set 2:

When I run with just a properties-content and no properties-file in my yaml file I do not get a properties-file section in the resulting xml.  I have made a change to not use get() anymore so there won't be a default of blank if someone provides properties-file but doesn't set it.  Test again with the latest and see if you still get an empty properties-file section.

Patch-set: 2
Label: Code-Review=0