Update patch set 2
Patch Set 2:

I had reworked the change a little, using a function to perform the work in the many places that it matters, and because it is very possible that the same thing could matter for new modules or options added in, just didn't get to upload the patchset yet.

The reqorked change also doesn't help, as given the if statement matching true previously, the XML element is still added.

Patch-set: 2
Label: Code-Review=0