Update patch set 3
Patch Set 3: I would prefer that you didn't merge this

I think it would be better to avoid having different code paths for test/non-test, so that one can actually use test to validate that a configuration will work (or be unchanged, or changed in a specific way, etc).

What are your thoughts on either of the following?
 a) move the authtoken out of the global config object and into the yaml
 b) support optionally using a configuration file in test mode

Patch-set: 3
Label: Verified=0
Label: Code-Review=-1
Label: Workflow=0