Fixing trigger-parameterized-builds
I tried using the trigger-parameterized-builds functionality of JJB and
I got the following error using the example given:
 projects.text = project_def['project']
 TypeError: string indices must be integers, not str
I took a look at the source and cannot figure out why the code is in a
for loop.  Projects to be triggered should be in a comma separated list
like you would do on the UI and you can only have one predefined
parameters section.  If you want to trigger other builds with different
parameters then you would just have another trigger-parameterized-builds
section in your yaml file.  So this change is to take the for loop out and
replace instances of project_def with data.

Change-Id: Ib6c5ffb18c213ccc1396edd1264edbf4c28ff651