Copy Artifact Plugin now specifies which build
The Copy Artifact Plugin has been extended to allow you to specify
which build to get.  There has been a dictionary created to
validate the value passed is correct and another dictionary
has been created to validate the permalink passed, if that option
is selected.  There are also parameters added for several of the
ways to select a build.

Change-Id: I71bbdcaada184736cc30e6390751aff1a864f9d5