Fixing svn-tags to work correctly
It was discovered that when using svn-tags functionality that when
you go to build the list is not populated.  If you just saved the job
in Jenkins then it would work correctly.  It was determined that the
<uuid> tag was missing and this was needed for the parameter to work.
In testing it was found that simply putting <uuid>1-1-1-1-1</uuid> into
the config would make the svn-tags functionality work like expected.
This is a one line change to add the uuid element set to 1-1-1-1-1

Change-Id: Iea1c27ebc9aa7c17f4f390c375989240c1941ce0