Update patch set 3
Patch Set 3:

I do believe it is more intuitive to have the leading dash because that is what every is used to for java options.  I left it off so you didn't have to quote, as it makes it uglier in my eyes.  I have submitted a change to make the dash necessary.

Patch-set: 3