gerrit: Make triggerOn* parameters really use boolean values.
Parameters triggerOn* are documented as boolean, however, using values like
true or false results in exception "TypeError: cannot serialize True (type
bool)".

Theses values have special meaning in yaml, and are not loaded as string in
python. It was hence required to use quotes to make it work.

Add a sample using thoses attributes as an example.

Change-Id: Iccbda8ebc0867e173fa7ccab78db16874ddb2a45