Use threads to update jobs in jenkins
Web applications such as jenkins are intended to handle
large numbers of requests in parallel, therefore use
threads to send update requests to allow many jobs to be
updated in parallel.

In order to support interrupts, ensure that after the
threads are started that the main process checks for a
finish result status and sleeps rather than joining all
worker threads.

Change-Id: Ie92be45d418b8076874b9eabca7e4ea8fb8fbad1