Support recursiving processing of paths
Multiple directories may be used to separate yaml files for different
groupings of jobs. Add new options to the parser for test and update
to recurse into the given path if it's a directory.

Provide an command line option to specify additional directories to be
excluded when recursive searching is enabled. Default is to only ignore
directories starting with a period - '.'.

Change function to expect a list of files or directories as an
argument, but support single specified files/dirs as well.

Change-Id: I236b364c268fd2bf5511a2c6d9a1c87914f3b086