Support recursive processing of paths
It is useful to be able to group yaml files under different paths within
the same repository to ease management of large numbers of files. Add
an option to the test and update subparsers to recursively process the
given path if it's a directory.

Include an additional option to exclude some paths/patterns from being
included in the search to allow a complex hierarchy of directories to be
referenced under a single path

Change load_files function to process list of files or directories.

Change-Id: I236b364c268fd2bf5511a2c6d9a1c87914f3b086