Error on duplicate job names being found
When parsing yaml files, any duplicate macros, job-templates, job-groups
or jobs should be considered an error by default, but provide a config
option for this to be ignored. Additionally check when creating jobs
from templates whether there is an existing job by the same name.

Also ensure that if the user wishes for duplicates to be ignored that
the last defined job replaces the earlier one internally before starting
any uploads to jenkins.

Change-Id: Id31ef33ae1b2469f7c3eeabe77420377978fa35e