Jobs not tied to a node can roam
Previously if you did not specify a node for a job to build on
then the job would be marked as tied to master.  The checkbox for
'Restrict where this project can be run' would be checked but
nothing entered into the textbox so it ended up tied to master.
This change adds an else to the check for 'node' and if 'node' is
not found then canRoam is set to true.  This is a two line change
to modules/general.py

Change-Id: Id7e40ba5ae2eaa92fd7678da7b3acefc1ba4ee32