'names' argument to update no longer required
Before the change to allow 'update' to take multiple jobs the
'names' argument was not required and passing just the yaml file
would create all the jobs inside it.  The 'update' change made
names required when it shouldn't be.  This is a one line change
to cmd.py to change nargs from '+' back to '?'

Change-Id: I65372d78ec678b9c29caad2ea573772287a7642f