Update patch set 3
Patch Set 3:

Yes, it is exposed in the "Build Triggers" section.

If the "Trigger builds remotely (e.g., from scripts)" checkbox  is enabled then a new textbox "Authentication Token" is showed.

It doesn't require a plug-in, in fact I just disabled all the plug-ins and restarted jenkins just to be sure.

Patch-set: 3
Label: Code-Review=0