RFC: Make reuse of builders/publishers inside other components easier.
Generation of XML markup from YAML markup is the job the _dispatch method
of the class Base. Unfortunately, components do not have access to their
module object, and even if their could, _dispatch would still be a
non-public method.

I've thought of several ways to offer that possibility:
- moving the core of _dispatch to an outer, public, function.
- extending the ModuleRegistry to register components or finding them when
requested.
- pass the calling module to component.

Leaving the logic of dispatch in the class Base seemed a better option
than moving it outside of any class.

Extending the ModuleRegistry is not ideal either: components can only
access it through their module, which is the problem we have to address
in the first place.

Which leaves the last solution, demonstrated in this change. Basically, it
means adding a module parameter to all components (which is a bit intrusive).

Interesting parts are in base.py and builders.py (at the end of the file).
Documentation is missing in that RFC.

Any suggestion on how to tackle that issue is welcome :)

Change-Id: Ie9d090817d0c2d464745b5634a22d3cea6a47ab1