Update patch set 1
Patch Set 1: I would prefer that you didn't merge this

All of the components currently get the YamlParser passed to them, which holds the ModuleRegistry as a member field.  They are actually so closely related that they probably don't need to be separate objects, but that's probably not important to change now.

Two alternate solutions:

1) Make a public API method on Builders that calls dispatch.

If the module registry had a mapping of name -> module (in addition to the current ordered list), it would be easy to get the 'builders' module from the registry.

You could then add a public method to the Builders class that invoked self._dispatch, and any builder would be able to use that.  It would work for both entrypoints builders and yaml macros.

2) Move _dispatch into ModuleRegistry

Make the assumption that every component registered via entrypoints or yaml macros can be invoked in the way it's currently being invoked via _dispatch.  Move Base._dispatch to ModuleRegistry.dispatch.  Call parser.registry.dispatch from any component to generate xml for any other component.

Now that I've thought about it a bit, I think that I like #2 the best.  Thoughts?

Patch-set: 1
Label: Code-Review=-1
Label: Workflow=0