Update patch set 1
Patch Set 1: I would prefer that you didn't merge this

I like this, but because it only uses entrypoints, it won't pick up YAML macros.  I think it would be worth addressing the issues in https://review.openstack.org/26051 first and basing the work on that so that the macros are universally available.

Thanks for doing this change first -- it really helped me understand what you wanted to do in 26051.  I think once we decide on a direction for that, this will be easy to update this one to use it.

Patch-set: 1
Label: Code-Review=-1
Label: Workflow=0