Update patch set 1
Patch Set 1:

The fact that #2 allows any component to use any other component clearly pleads for it.

It has the caveat to require components to provide both component_type and component_list_type when calling dispatch, which seems to me like implementation details of modules.

I think it can be improved by mixing the idea of mapping a name to a module:
- add two class level members (component_type and component_list_type) in modules which have components, similar to sequence.
- make ModuleRegistry map component_type to modules,
- refactor ModuleRegistry.dispatch to take only component_type, and deduce component_list_type from that.

What do you think?

Patch-set: 1
Label: Code-Review=0