Update patch set 1
Patch Set 1:

Why is generating the XML bad?

It already does not write the XML if the job isn't in the list of job names.  This adds some complexity just to avoid extra computation during the unusual situation of just wanting a few jobs (the normal mode of operation is that all jobs should be generated).

I wonder if the code complexity is worth it?

Patch-set: 1
Label: Code-Review=0
Label: Workflow=0