Update patch set 1
Patch Set 1:

On our setup we have a total of 1680 jobs, 1613 are for MediaWiki extensions

The 67 other jobs are updated more often.

So we end up only regenerating the jobs we are interested in (for example all the puppet related jobs).  Waiting for the 1613 xml generations is a bit troublesome.

I am not sure how to optimize that in a better way. One possibility would be to split our configurations in two directories and find a way to have some common directory to share macros between them.

Patch-set: 1
Label: Code-Review=0