Update patch set 6
Patch Set 6: I would prefer that you didn't merge this

This is great.  The original 'coverage' publisher in JJB was based far too much on what openstack happened to be doing with it at the time, so I'm happy to get rid of it, and your implementation looks great.

However, I think we should do a deprecation/removal cycle with this to avoid suddenly breaking current users.

What if you called the new publisher 'cobertura' (I don't think that would be incorrect), and marked the current coverage publisher as deprecated.  We'll issue a release with both included, then remove 'coverage' later.  We can then decide whether we want to alias cobertura to coverage or just leave it as is.

I don't think we've deprecated something like this in JJB before; I expect that would involve a documentation update, and perhaps issuing a warning log message if coverage is used.

Would you mind adjusting your patchset to accomodate both publishers?  If you feel like performing the deprecation work too, that would be appreciated.

Thanks.

Patch-set: 6
Label: Code-Review=-1
Label: Workflow=0