Pass svn revision to triggered jobs
This change adds the ability to pass the subversion revisions that
were used in the build to the downstream builds.

Added passing current-build parameters to downstream jobs
triggered in the post build actions, to keep builders and publishers
in sync.

Noticed arg type for git-revision was incorrect and changed it to
bool from str.

Change-Id: I029c1d88b23e08ac2546a92018fef06c22c357b1