Update patch set 2
Patch Set 2: Looks good to me, but someone else must approve

LGTM.  When reading the documentation, I wasn't quite sure why jmeter and junit were listed in the reports section twice.  After re-reading it, I came to the conclusion that the first two were to pick up special files, and the second two were to additionally use the default files.  Reading the code suggested that interpretation is correct.

It might be helpful to explain the example a little bit, or split it into multiple examples.

Patch-set: 2
Label: Code-Review=+1
Label: Workflow=0