Update patch set 2
Patch Set 2: I would prefer that you didn't merge this

I agree with jeblair, the example yaml is a little confusing and should probably be split into two examples. Also, what do you think about adding an else block to your if elif checks for junit and jmeter that throws an exception to fail hard? It might be hard to understand why this isn't working properly if a key is typoed or a different key is used, but failing hard should hopefully make that apparent.

Patch-set: 2
Label: Code-Review=-1
Label: Workflow=0