Update patch set 1
Patch Set 1:

I feel like maintaining backward compatibility somehow would be nice, but at the same time am having a hard time thinking up a way to do that cleanly.

We could add a different builder that changes the key from projectName to project. Or we could add a flag to the existing builder that sets a version that we can switch off of. I don't like either of these options.

JJB is a simple tool and I don't want to overcomplicate it and keeping up with the plugins is probably more important than maintaining backward compatibility in this particular case. I will have to think about this more.

Patch-set: 1
Label: Code-Review=0
Label: Workflow=0