Update patch set 6
Patch Set 6:

All of the credentials are hidden away in the global config.  It would be nice if JJB could set the global config somehow.  There is only room for one user and password so he would have to have access to all the rooms.  There isn't even a space for the room password in the job config but there is in the global config where you can setup rooms to connect to automatically, as well as an alias for the user and some other config

Patch-set: 6