Update patch set 1
Patch Set 1:

I notice there is a field for a user password, but you don't fill it in.

If you did, would that password be updated?  (In other words, is the 'you must update the password manually in the web UI' bit a choice you made when implementing this, or does it actually not work to upload the password in the XML config?)

If it does work, you could use the model employed by the hipchat_notif module to store the password in the global config file, so that it is kept separate from the YAML.

Patch-set: 1
Reviewer: Gerrit User 1 <1@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=0
Label: Workflow=0