Update patch set 1
Patch Set 1: (1 inline comment)

There is some talk of adding a new feature to JJB that would allow inlining the contents of another file into the yaml. This may make it semi reasonable for JJB to set passwords. See inline comments for details and maybe hop on IRC to chat about this feature.

Patch-set: 1